create refs database
