+++ lib/models/app_model.dart
@@ -8,8 +8,8 @@ import 'package:flutter/material.dart';
 import 'package:flutter_settings_screens/flutter_settings_screens.dart';
 import 'package:fuzzywuzzy/fuzzywuzzy.dart';
 import 'package:hive/hive.dart';
-import 'package:otzaria/data/data.dart';
-import 'package:otzaria/data/file_system_data_provider.dart';
+import 'package:otzaria/data/repository/data.dart';
+import 'package:otzaria/data/data_providers/file_system_data_provider.dart';
 import 'package:otzaria/models/bookmark.dart';
 import 'package:otzaria/models/books.dart';
 import 'package:otzaria/models/library.dart';
