+++ lib/models/books.dart
@@ -1,5 +1,5 @@
-import 'package:otzaria/data/data.dart';
-import 'package:otzaria/data/file_system_data_provider.dart';
+import 'package:otzaria/data/repository/data.dart';
+import 'package:otzaria/data/data_providers/file_system_data_provider.dart';
 import 'package:otzaria/models/links.dart';
 import 'dart:isolate';
 //import 'package:pdfrx/pdfrx.dart';
@@ -115,7 +115,7 @@ class TextBook extends Book {
 ///represents an entry in table of content , which is a node in a hirarchial tree of topics.
 ///every entry has its 'level' in the tree, and an index of the line in the book that it is refers to
 class TocEntry {
-  final String text;
+  String text;
   final int index;
   final int level;
   List<TocEntry> children = [];
