+++ lib/models/links.dart
@@ -1,9 +1,9 @@
 /* represents links between two books in the library*/
 
-import 'package:otzaria/data/file_system_data_provider.dart';
+import 'package:otzaria/data/data_providers/file_system_data_provider.dart';
 import 'package:otzaria/models/books.dart';
 import 'dart:isolate';
-import 'package:otzaria/data/data.dart';
+import 'package:otzaria/data/repository/data.dart';
 import 'package:otzaria/utils/text_manipulation.dart' as utils;
 
 /// Represents a link between two books in the library.
