+++ pubspec.yaml
@@ -51,7 +51,8 @@ dependencies:
     sdk: flutter
   flutter_settings_screens: ^0.3.4
   hive: ^4.0.0-dev.2
-  isar_flutter_libs: ^4.0.0-dev.13
+  isar_flutter_libs: ^4.0.0-dev.14
+  isar: ^4.0.0-dev.14
   msix: ^3.16.7
   path_provider: ^2.0.15
   html: ^0.15.1
@@ -103,6 +104,7 @@ dev_dependencies:
   # rules and activating additional ones.
   flutter_lints: ^2.0.0
   test: ^1.25.2
+  build_runner: ^2.4.11
 
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
